V3 (#3)
* time

* miprove time

* test support no time limit

* rename U64 to uint

* rename TIME_BEGIN to TIME RESET

* case custom timeout

* run on mac

* int128_t

* fix

* assert_dbg

* more assert

* bettertesting and debug

* macro bit shift

* assert log on prod

* mid commit

* better uint macros

* dprintf

* tprintf

* isa native print foramt

* fix time

* remove deprecated usleep

+++ test.h
@@ -7,8 +7,10 @@
 #include <stdlib.h>
 #include <signal.h>
 #include <stdarg.h>
+#include <time.h>
 
 #include "assert.h"
+#include "uint.h"
 
 #define TEST_LIB printf("\n%s\t\t", __func__);
 
@@ -17,7 +19,7 @@
         printf("\n\t%s\t\t", __func__); \
         bool __is_main_process = true;  \
         bool __test_memory = true;      \
-        int __tag = 0;
+        uint64_t __tag = 0;
 
 #define TEST_FN_CLOSE               \
         if(__test_memory)           \
@@ -26,17 +28,17 @@
             exit(EXIT_SUCCESS);     \
     }
 
-void test_log_error(int __tag, int line, char const func[], char format[], ...)
+void test_log_error(uint64_t __tag, uint64_t line, char const func[], char format[], ...)
 {
     va_list args;
     va_start(args, format);
-    printf("\n\n\tERROR TEST\t| l:%d | %s %d | ", line, func, __tag);
+    printf("\n\n\tERROR TEST\t| l: " U64P() " | %s " U64P() " | ", line, func, __tag);
     vprintf(format, args);
     printf("\n\n");
     exit(EXIT_FAILURE);
 }
 
-pid_t fork_assert(int __tag, int line, char const func[], char fork_tag[])
+pid_t fork_assert(uint64_t __tag, uint64_t line, char const func[], char fork_tag[])
 {
     pid_t pid = fork();
     if(pid < 0)
@@ -46,51 +48,69 @@ pid_t fork_assert(int __tag, int line, char const func[], char fork_tag[])
     return pid;
 }
 
+pid_t waitpid_assert(uint64_t __tag, uint64_t line, char const func[], pid_t pid, int *status)
+{
+    pid_t pid_return = waitpid(pid, status, 0);
+    if(pid_return <= 0)
+    {
+        test_log_error(__tag, line, func, "WAITPID RETURNED %d", pid_return);
+    }
+    return pid_return;
+}
+
 // returns true if main process
-bool start_case(int __tag, int line, char const func[], bool show)
+bool start_case(uint64_t __tag, uint64_t line, char const func[], bool show, uint64_t timeout_ms)
 {
     if(show)
-        printf("\n\t\t%s %2d\t\t", func, __tag);
+        printf("\n\t\t%s " U64P(2) "\t\t", func, __tag);
 
     pid_t pid = fork_assert(__tag, line, func, "TEST");
     if(pid)
     {
         int status;
-        waitpid(pid, &status, 0);
+        waitpid_assert(__tag, line, func, pid, &status);
         assert(status == EXIT_SUCCESS);
         return true;
     }
 
-    int pid_test = fork_assert(__tag, line, func, "TEST");
+    pid_t pid_test = fork_assert(__tag, line, func, "TEST");
     if(pid_test == 0)
         return false;
 
-    pid_t pid_timeout = fork_assert(__tag, line, func, "TIMEOUT");
-    if(pid_timeout == 0)
-    {
-        usleep(TEST_CASE_TIMEOUT_MS * 1000);
-        exit(EXIT_SUCCESS);
-    }
-
     int status;
-    pid_t pid_return = waitpid(0, &status, 0);
-    if(pid_return < 0)
+    if(timeout_ms)
     {
-        test_log_error(__tag, line, func, "WAITPID RETURNED %d", pid_return);
+        pid_t pid_timeout = fork_assert(__tag, line, func, "TIMEOUT");
+        if(pid_timeout == 0)
+        {
+            struct timespec spec = (struct timespec)
+            {
+                .tv_sec = (long)(timeout_ms / 1000),
+                .tv_nsec = (long)((timeout_ms % 1000) * 1000000)
+            };
+            nanosleep(&spec, NULL);
+            exit(EXIT_SUCCESS);
+        }
+
+        pid_t pid_return = waitpid_assert(__tag, line, func, 0, &status);
+        if(pid_return == pid_timeout)
+        {
+            kill(pid_test, SIGKILL);
+            test_log_error(__tag, line, func, "TEST TIMEOUT");
+        }
+
+        if(pid_return != pid_test)
+        {
+            test_log_error(__tag, line, func, "INVALID PID CAUGHT %d", pid_return);
+        }
+
+        kill(pid_timeout, SIGKILL);
+    }
+    else
+    {
+        waitpid_assert(__tag, line, func, pid_test, &status);
     }
 
-    if(pid_return == pid_timeout)
-    {
-        kill(pid_test, SIGKILL);
-        test_log_error(__tag, line, func, "TEST TIMEOUT");
-    }
-
-    if(pid_return != pid_test)
-    {
-        test_log_error(__tag, line, func, "INVALID PID CAUGHT %d", pid_return);
-    }
-
-    kill(pid_timeout, SIGKILL);
     if(status != EXIT_SUCCESS)
     {
         test_log_error(__tag, line, func, "ERROR IN TEST EXECUTION ");
@@ -98,21 +118,25 @@ bool start_case(int __tag, int line, char const func[], bool show)
     exit(EXIT_SUCCESS);
 }
 
-#define TEST_CASE_OPEN(TAG)                                                 \
-    if(__is_main_process)                                                   \
-    {                                                                       \
-        __tag = (int)(TAG);                                                 \
-        __is_main_process = start_case(__tag, __LINE__, __func__, show);    \
-        if(!__is_main_process)                                              \
+
+
+#define TEST_CASE_OPEN_TIMEOUT(TAG, TIMEOUT)                                        \
+    if(__is_main_process)                                                           \
+    {                                                                               \
+        __tag = (uint64_t)(TAG);                                                    \
+        __is_main_process = start_case(__tag, __LINE__, __func__, show, TIMEOUT);   \
+        if(!__is_main_process)                                                      \
         {
+
+#define TEST_CASE_OPEN(TAG) TEST_CASE_OPEN_TIMEOUT(TAG, TEST_CASE_TIMEOUT_MS)
 
 #define TEST_CASE_CLOSE \
         }               \
     }
 
-pid_t start_revert(int __tag, int line, char const func[])
+pid_t start_revert(uint64_t __tag, uint64_t line, char const func[])
 {
-    int pid = fork();
+    pid_t pid = fork();
     if(pid < 0)
     {
         test_log_error(__tag, line, func, "ERROR FORKING");
@@ -120,10 +144,7 @@ pid_t start_revert(int __tag, int line, char const func[])
     if(pid)
     {
         int status;
-        if(waitpid(pid, &status, 0) < 0)
-        {
-            test_log_error(__tag, line, func, "WAITPID RETURNED INVALID");
-        }
+        waitpid_assert(__tag, line, func, pid, &status);
         if(status == EXIT_SUCCESS)
         {
             test_log_error(__tag, line, func, "TEST EXPECTED TO REVERT BUT DIDN'T");
@@ -139,7 +160,8 @@ pid_t start_revert(int __tag, int line, char const func[])
             printf("\n\n\tERROR REDIRECTING STD BUFFERS\n\n");
             exit(EXIT_SUCCESS);
         }
-        usleep(0);
+        struct timespec spec = (struct timespec) {0};
+        nanosleep(&spec, NULL);
     }
     return pid;
 }
